# Split out declared dependencies