More perf work. `BuildTarget.DeclaredDependencies()` shows up as 2-3% of CPU, which seems too much; this makes it inline to basically nothing.
Before:
```
INFO: Run 1 of 5
INFO: Complete in 27.83s, using 9933692 KB
INFO: Run 2 of 5
INFO: Complete in 24.55s, using 12287520 KB
INFO: Run 3 of 5
INFO: Complete in 23.06s, using 13040620 KB
INFO: Run 4 of 5
INFO: Complete in 23.29s, using 13311044 KB
INFO: Run 5 of 5
INFO: Complete in 23.23s, using 13587596 KB
INFO: Complete, median time: 23.29s, median mem: 13040620.00 KB
```
After:
```
INFO: Run 1 of 5
INFO: Complete in 22.96s, using 13285284 KB
INFO: Run 2 of 5
INFO: Complete in 22.80s, using 14015176 KB
INFO: Run 3 of 5
INFO: Complete in 21.29s, using 11657828 KB
INFO: Run 4 of 5
INFO: Complete in 20.67s, using 12313460 KB
INFO: Run 5 of 5
INFO: Complete in 21.76s, using 14261576 KB
INFO: Complete, median time: 21.76s, median mem: 13285284.00 KB
```