@peterebden 2024-05-15
Hmmmmm. There is some non-determinism in how dependencies are populated on the target (stemming from cases where `srcs` etc are dicts).
I think this might be one to come back to once asp has native ordered dicts (for which I really want rangefunc)